🐛 fix: Fix image upload error
