+++ src/app/chat/(mobile)/features/ChatInput/Files.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
+import FileList from 'src/components/FileList';
 
-import FileList from '@/app/chat/components/FileList';
 import { useFileStore } from '@/store/file';
 
 const Files = memo(() => {
