+++ src/components/FileList/Lightbox.tsx
@@ -1,8 +1,6 @@
 import { Image } from 'antd';
 import { PropsWithChildren, memo } from 'react';
 
-import { filesSelectors, useFileStore } from '@/store/file';
-
 const { PreviewGroup } = Image;
 
 interface LightBoxProps extends PropsWithChildren {
@@ -10,13 +8,10 @@ interface LightBoxProps extends PropsWithChildren {
 }
 
 const LightBox = memo<LightBoxProps>(({ items, children }) => {
-  const list = useFileStore(filesSelectors.getImageUrlByList(items));
-
-  if (list.length === 1) return children;
+  if (items.length === 1) return children;
 
   return (
     <PreviewGroup
-      items={list}
       preview={{
         styles: { mask: { backdropFilter: 'blur(2px)' } },
       }}
